# reduce allocs in populateDeviceField