--- a/pkg/metrics/series.go
+++ b/pkg/metrics/series.go
@@ -86,14 +86,13 @@ func (series Series) Marshal() ([]byte, error) {
 
 // populateDeviceField removes any `device:` tag in the series tags and uses the value to
 // populate the Serie.Device field
-// Mutates the `series` slice in place
 //FIXME(olivier): remove this as soon as the v1 API can handle `device` as a regular tag
 func populateDeviceField(series Series) {
 	for _, serie := range series {
 		// make a copy of the tags array. Otherwise the underlying array won't have
 		// the device tag for the Nth iteration (N>1), and the device field will
 		// be lost
-		var filteredTags []string
+		filteredTags := make([]string, 0, len(serie.Tags))
 
 		for _, tag := range serie.Tags {
 			if strings.HasPrefix(tag, "device:") {