--- a/pkg/metrics/series_benchmark_test.go
+++ b/pkg/metrics/series_benchmark_test.go
@@ -20,6 +20,7 @@ func benchmarkPopulateDeviceField(numberOfTags int, b *testing.B) {
 	series := []*Serie{serie}
 
 	for n := 0; n < b.N; n++ {
+		serie.Tags = tags
 		populateDeviceField(series)
 	}
 }