@olivielpeau 2019-02-12
@arbll could you add a release note for this PR as well? (and/or couple it with https://github.com/DataDog/datadog-agent/pull/3006)